+++ server/src/trace.rs
@@ -18,7 +18,7 @@ pub fn init_tracing(config: &crate::config::Config) -> Option<tracing_chrome::Fl
 
     match config.opts.trace {
         crate::config::Tracing::Stdout => {
-            let terminal_layer = tracing_subscriber::fmt::Layer::default().compact();
+            let terminal_layer = tracing_subscriber::fmt::Layer::default();
             tracing_registry.with(terminal_layer).init();
         }
         crate::config::Tracing::Chrome => {